Created subclasses of menu controller for pause and main menus
