--- Cosmic Quest - Order and Chaos/Assets/Collections/MK Glow/MKGlow/Scripts/LWRP.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f303c366664628b43a4e5aa5c94d2fce
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
