+++ Cosmic Quest - Order and Chaos/Assets/Scenes/MenuStaging.unity
@@ -2641,8 +2641,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 9eec87cb399e585458b08f3cc02a8851, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  ActiveMenu: {fileID: 1635574747}
-  LobbyConfirmButton: {fileID: 339216149}
+  activeMenu: {fileID: 1635574747}
+  lobbyConfirmButton: {fileID: 339216149}
   playerJoinAreas:
   - {fileID: 1245533008}
   - {fileID: 337300843}
