+++ Cosmic Quest - Order and Chaos/Assets/Scripts/Menu/MainMenuController.cs.meta~HEAD
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 768079e7ae373884899e66d1b7181a42
+guid: 9eec87cb399e585458b08f3cc02a8851
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
